Break each screenshot definition into its own file
